Moved progresdsindicator package
