+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListActivity.kt
@@ -28,6 +28,7 @@ import com.numero.material_gallery.components.fab.FabActivity
 import com.numero.material_gallery.components.image.ShapeableImageViewActivity
 import com.numero.material_gallery.components.navigationdrawer.NavigationDrawerActivity
 import com.numero.material_gallery.components.picker.DatePickerActivity
+import com.numero.material_gallery.components.progressindicator.ProgressBarActivity
 import com.numero.material_gallery.components.selection.CheckboxActivity
 import com.numero.material_gallery.components.selection.RadioButtonActivity
 import com.numero.material_gallery.components.selection.SwitchActivity
